moved non-exec cmd/util package to cliutil
