+++ cliutil/utils.go
@@ -1,4 +1,4 @@
-package utils
+package cliutil
 
 import (
 	"crypto/ecdsa"
