+++ cliutil/utils_test.go
@@ -1,4 +1,4 @@
-package utils
+package cliutil
 
 import (
 	"encoding/hex"
